Move CharacterPoolInterface to models/game/pools/interfaces
