+++ frontend/models/game/Game.php
@@ -14,8 +14,8 @@ use frontend\exceptions\CannotStartWithoutCharacterException;
 use frontend\exceptions\FinishedGameException;
 use frontend\exceptions\NotStartedGameException;
 use frontend\models\game\base\BeeInterface;
-use frontend\models\game\base\CharacterPoolInterface;
 use frontend\models\game\base\HoneyPoolInterface;
+use frontend\models\game\pools\interfaces\CharacterPoolInterface;
 use frontend\models\game\characters\interfaces\PlayerInterface;
 
 class Game implements GameInterface
