+++ tests/game/inside/CharacterPoolTest.php
@@ -10,7 +10,7 @@ namespace tests\game\inside;
 
 
 use frontend\models\game\base\BeeInterface;
-use frontend\models\game\base\CharacterPoolInterface;
+use frontend\models\game\pools\interfaces\CharacterPoolInterface;
 use frontend\models\game\characters\Drone;
 use frontend\models\game\characters\Healer;
 use frontend\models\game\characters\Player;
